refactor(contests): split the scorer into service and queries
Pure refactor. No behaviour change, no scoring change; CONTEST_SCORE_CODE_VERSION
stays at 5.

contest-score.service.ts had grown past 2,300 lines and an earlier review flagged
its size as a contributing factor to the Turbopack chunk-name collision. The
data-access layer now lives in contest-score.queries.ts: every Postgres and
ClickHouse query, the gate helpers, the id-list and lookup-table builders, and
the shared row types.

The seam is where the module stops knowing about scoring. Queries knows how to
ask the two stores things and nothing about weights, normalization, ranking,
snapshots or config; the service keeps all of that and no longer contains SQL.
The ClickHouse source builders moved with it -- they were inline template
literals in two functions, so leaving them behind would have kept query text on
the wrong side of the line.

The dependency runs one way. Queries imports nothing from the service, so the
shared types, DEFAULT_STATUSES and ContestScoringError live at the bottom of the
graph; the service re-exports the three names external callers already import
from it.

Prisma is now imported as a type in the service -- it is only used for
Prisma.InputJsonValue since the raw-SQL helpers moved out, and a value import
with no value references crashes @typescript-eslint/consistent-type-imports.
